==> analysis/05-fig02-ts.R <==
library(dplyr)
library(ggplot2)
source("analysis/theme.R")

hbll <- readRDS("data-generated/index-hbll-geo-clean.rds")
syn <- readRDS("data-generated/index-syn-geo-clean.rds")
index <- bind_rows(hbll, syn)

index <- mutate(index, species_common_name = gsub(
  "rougheye/blackspotted rockfish complex",
  "rougheye/blackspotted rockfish", species_common_name
)) |>
  filter(!(year == 2014 & survey_abbrev == "SYN WCHG"))

metrics <- readRDS("data-generated/metrics-wide2.rds")

prop <- metrics |>
  select(
    prop_mpa, species_common_name,
    survey_abbrev
  ) |>
  distinct() |>
  mutate(species_common_name = tolower(species_common_name))

index$prop_mpa <- NULL
index <- left_join(index, prop)

top_prop <- prop |>
  filter(prop_mpa > 0.15) |>
  filter(!survey_abbrev %in% c("SYN QCS, SYN HS")) |>
  group_by(survey_abbrev) |>
  arrange(survey_abbrev, -prop_mpa) |>
  mutate(prop_mpa = round(prop_mpa, 2))

top_prop |> as.data.frame()

dodge_width <- 1

colour_pal <- c("gray50", restricted_cols)

fig2_keep <- readr::read_csv("data-raw/selected-spp.csv") |>
  select(-prop_mpa, -notes) |>
  filter(include == 1) |>
  mutate(species_common_name_lower = tolower(species_common_name))
nrow(fig2_keep)

ind <- index |>
  mutate(species_common_name_lower = tolower(species_common_name)) |>
  semi_join(fig2_keep, by = join_by(survey_abbrev, species_common_name_lower))

labeller_fn <- function(labels, multi_line = TRUE) {
  labels <- lapply(labels, function(x) {
    gsub("[a-zA-Z ,]+-", "", as.character(x))
  })
  if (multi_line) {
    labels
  } else {
    collapse_labels_lines(labels)
  }
}

cols <- c(RColorBrewer::brewer.pal(3L, "Set2"), "grey50")
dd <- ind |>
  mutate(spp_survey = species_common_name) |>
  mutate(spp_survey = gsub("HBLL OUT N", "HBLL", spp_survey)) |>
  mutate(spp_survey = gsub("rougheye/blackspotted rockfish", "rougheye/blackspotted", spp_survey)) |>
  filter(type %in% c("Status quo", "Restricted and shrunk")) |>
  group_by(spp_survey, type, survey_abbrev) |>
  mutate(
    colour_var =
      ifelse(
        type == "Status quo",
        paste0("Status quo ", survey_abbrev),
        paste0("Restricted")
      )
  ) |>
  mutate(colour_var = factor(colour_var,
    levels = c(
      "Status quo",
      "Status quo SYN WCHG",
      "Status quo HBLL OUT N",
      # "Status quo SYN QCS, SYN HS"
      "Status quo SYN HS",
      "Status quo SYN QCS",
      "Restricted"
    )
  )) |>
  mutate(survey_abbrev = factor(survey_abbrev,
    levels = c(
      "SYN WCHG",
      "HBLL OUT N",
      "SYN HS",
      "SYN QCS"
    )
  )) |>
  mutate(lwr = lwr / exp(mean(log(est), na.rm = TRUE))) |>
  mutate(upr = upr / exp(mean(log(est), na.rm = TRUE))) |>
  mutate(est = est / exp(mean(log(est), na.rm = TRUE))) |>
  mutate(AFFECTED = ifelse(upr > 4 * max(est, na.rm = TRUE), "YES", "NO")) |>
  mutate(upr = ifelse(upr > 4 * max(est, na.rm = TRUE), 4 * max(est, na.rm = TRUE), upr)) |>
  mutate(spp_survey = gsub("Rougheye/Blackspotted Rockfish", "Rougheye/Black. Rockfish", spp_survey)) |>
  arrange(survey_abbrev, -prop_mpa, species_common_name)
filter(dd, AFFECTED == "YES")

g <- dd |> ggplot(aes(year, est,
  ymin = lwr, ymax = upr,
  colour = colour_var
)) +
  geom_linerange(position = position_dodge(width = dodge_width)) +
  geom_point(position = position_dodge(width = dodge_width), pch = 21, size = 1.8) +
  geom_point(position = position_dodge(width = dodge_width), pch = 20, size = 1.8, alpha = 0.2) +
  coord_cartesian(
    expand = FALSE,
    xlim = range(index$year) + c(-0.5, 0.5),
    ylim = c(0, NA)
  ) +
  labs(x = "Year", colour = " ", fill = " ", linetype = " ") +
  scale_colour_manual(values = cols) +
  facet_wrap(~ forcats::fct_inorder(stringr::str_to_title(paste(survey_abbrev, spp_survey, sep = "-"))), scales = "free_y", ncol = 5, labeller = labeller_fn) +
  ylab("Relative abundance or biomass") +
  labs(x = "Year", colour = "Index type", fill = "Index type", linetype = "Index type") +
  theme(
    strip.text = element_text(colour = "black"),
    legend.position = "top", axis.text.y = element_text(size = 8)
  ) +
  geom_smooth(method = "gam", se = FALSE, alpha = 0.1, formula = y ~ s(x, k = 8), method.args = list(family = Gamma(link = "log")), lwd = 0.8, mapping = aes(fill = colour_var)) +
  scale_fill_manual(values = cols)
g

g <- g + tagger::tag_facets(tag_prefix = "(", position = list(x = 0.08, y = 0.89), tag = "panel")

ggsave("figs/index-geo-restricted-highlights.pdf", width = 9.5, height = 5.25)
ggsave("figs/index-geo-restricted-highlights.png", width = 9.5, height = 5.25)
